+++ CHANGELOG.markdown
@@ -3,6 +3,12 @@
 
 ## Version 0.9.3 - *Not Yet Released*
 
+ * [Aphid.UI.View] Implemented automatic Event observer registration (and
+   de-registration) for any View class that implements a method to handle the
+   event (i.e. "handleClickEvent", "handleMouseMoveEvent", etc). Events will
+   be observed when the View appears and observers will be removed when the
+   View disappears.
+
  * [Aphid.UI.MatrixView] Added two new configuration parameters for
    controlling multiple selection and drag selection support. Drag selection
    requires multiple selection to be enabled.